+++ artemis.egg-info/requires.txt
@@ -2,4 +2,3 @@ numpy
 matplotlib
 seaborn
 pandas
-pybind11
