+++ artemis.egg-info/top_level.txt
@@ -1,2 +1,1 @@
-_test
 artemis
