--- artemis/allostery/allostery.py
@@ -1,6 +1,0 @@
-from artemis.allostery.draw import draw
-from artemis.allostery.analysis import analysis
-
-
-
-def allostery(args):
