--- test.py
@@ -1,5 +1,0 @@
-import ctypes
-
-test = ctypes.CDLL('denoise')
-
-test.main()
